+++ src/menu/shell/shell.h
@@ -14,19 +14,20 @@
  * You should have received a copy of the GNU General Public License along
  * with swanshell. If not, see <https://www.gnu.org/licenses/>.
  */
- 
+
 #ifndef SHELL_H_
 #define SHELL_H_
- 
+
 #include <stdbool.h>
 #include <stdint.h>
 #include <wonderful.h>
 
 void shell_init(void);
-void shell_tick(void);
+bool shell_tick(void);
 
 enum {
     SHELL_RET_IDLE = 0,
+    SHELL_RET_REFRESH_UI,
     SHELL_RET_LAUNCH_IN_PSRAM,
 };