+++ docs/content/user/changelog.md
@@ -5,6 +5,7 @@ weight: 100
 
 ## swanshell 1.1.0 (Eventually)
 
+- Added: Enhanced USB shell (`help`, `reboot`, `upload`, `launch path` commands).
 - Added: Saving `/ram0` contents when loading `.fx` programs directly.
 - Added: Status icons - the cartridge's status is now communicated visually in the bottom right corner.
 - Added: The option to hide save files.